backend/requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
django-filter==24.3
openpyxl==3.1.5
uvicorn==0.30.6
gunicorn==22.0.0
python-dotenv==1.0.1
django-cors-headers==4.4.0
whitenoise==6.7.0
